add optional UART for PowerPC CPU
this modules effectively redirects all stdout either from EE or IOP to the PowerPC UART.

it cleanly supports multithreading, and it will not suffer from data loss/truncation like udptty

this module is safe to arbitrarly load. since it is coded to abort its startup if console is not both real hardware and a DEKCARD Unit.